Move write_json() to public API
This function was not initially intended for external use, but that's
happened over the years along with improvements and breaking changes to
the function itself.

Moving it to the public API indicates future changes will be tracked in
the changelog and will affect Augur's semantic versioning scheme.
